+++ statuspage/extras/forms/bulk_import.py
@@ -21,7 +21,7 @@ class WebhookImportForm(CSVModelForm):
     class Meta:
         model = Webhook
         fields = (
-            'name', 'enabled', 'content_types', 'type_create', 'type_update', 'type_delete',
+            'name', 'subscriber', 'enabled', 'content_types', 'type_create', 'type_update', 'type_delete',
             'payload_url', 'http_method', 'http_content_type', 'additional_headers', 'body_template',
             'secret', 'ssl_verification', 'ca_file_path'
         )
