+++ statuspage/extras/forms/model_forms.py
@@ -8,6 +8,7 @@ from utilities.forms.fields import ContentTypeMultipleChoiceField
 
 __all__ = (
     'WebhookForm',
+    'PublicWebhookForm',
 )
 
 
@@ -18,7 +19,7 @@ class WebhookForm(TailwindMixin, forms.ModelForm):
     )
 
     fieldsets = (
-        ('Webhook', ('name', 'content_types', 'enabled')),
+        ('Webhook', ('name', 'subscriber', 'content_types', 'enabled')),
         ('Events', ('type_create', 'type_update', 'type_delete')),
         ('HTTP Request', (
             'payload_url', 'http_method', 'http_content_type', 'additional_headers', 'body_template', 'secret',
@@ -39,4 +40,30 @@ class WebhookForm(TailwindMixin, forms.ModelForm):
             'additional_headers': forms.Textarea(attrs={'class': 'font-mono'}),
             'body_template': forms.Textarea(attrs={'class': 'font-mono'}),
             'conditions': forms.Textarea(attrs={'class': 'font-mono'}),
+        }
+
+
+class PublicWebhookForm(TailwindMixin, forms.ModelForm):
+    fieldsets = (
+        ('Webhook', ('name',)),
+        ('Events', ('type_create', 'type_update', 'type_delete')),
+        ('HTTP Request', (
+            'payload_url', 'http_method', 'http_content_type', 'additional_headers', 'body_template', 'secret',
+        )),
+        ('SSL', ('ssl_verification',)),
+    )
+
+    class Meta:
+        model = Webhook
+        fields = ('name', 'type_create', 'type_update', 'type_delete', 'payload_url', 'http_method', 'conditions',
+                  'http_content_type', 'additional_headers', 'body_template', 'secret', 'ssl_verification')
+        labels = {
+            'type_create': 'Creations',
+            'type_update': 'Updates',
+            'type_delete': 'Deletions',
+        }
+        widgets = {
+            'additional_headers': forms.Textarea(attrs={'class': 'font-mono'}),
+            'body_template': forms.Textarea(attrs={'class': 'font-mono'}),
+            'conditions': forms.HiddenInput(),
         }
