+++ statuspage/templates/home/subscribers/manage/webhook_create.html
@@ -10,21 +10,25 @@
   <div>
     <a href="{% url 'home' %}" class="bg-indigo-500 hover:bg-indigo-600 text-white shadow px-4 py-2 rounded-md" role="button">Home</a>
   </div>
+  <div>
+    <a href="{% url 'subscriber_manage_webhook_list' management_key=subscriber.management_key %}" class="bg-indigo-500 hover:bg-indigo-600 text-white shadow px-4 py-2 rounded-md" role="button">Webhooks</a>
+  </div>
 {% endblock %}
 
 {% block home_content %}
 <div class="flex flex-col space-y-8">
   <div class="space-y-4">
-    <div class="text-2xl">Manage Subscription</div>
-    <form action="" method="post">
-      {% csrf_token %}
-      <div class="flex flex-col space-y-2">
-        {% render_form form %}
-        <button type="submit" class="bg-indigo-500 hover:bg-indigo-600 text-white shadow px-4 py-2 rounded-md">Update</button>
-      </div>
-    </form>
+    <div class="flex flex-row justify-between">
+      <div class="text-2xl">Webhooks</div>
+    </div>
     <div>
-      <a href="{% url 'subscriber_manage_webhook_list' management_key=subscriber.management_key %}" class="bg-indigo-500 hover:bg-indigo-600 text-white shadow px-4 py-2 rounded-md" role="button">Webhooks</a>
+      <form action="" method="post">
+        {% csrf_token %}
+        <div class="flex flex-col space-y-2">
+          {% render_form form %}
+          <button type="submit" class="bg-indigo-500 hover:bg-indigo-600 text-white shadow px-4 py-2 rounded-md">Create Webhook</button>
+        </div>
+      </form>
     </div>
   </div>
 </div>
