+++ statuspage/templates/home/subscribers/manage/webhook_delete.html
@@ -1,0 +1,35 @@
+{% extends 'base/home.html' %}
+{% load helpers %}
+{% load form_helpers %}
+{% load buttons %}
+
+{% block controls %}
+  <div>
+    <a href="{% url 'subscriber_unsubscribe' management_key=subscriber.management_key %}" class="bg-indigo-500 hover:bg-indigo-600 text-white shadow px-4 py-2 rounded-md" role="button">Unsubscribe</a>
+  </div>
+  <div>
+    <a href="{% url 'home' %}" class="bg-indigo-500 hover:bg-indigo-600 text-white shadow px-4 py-2 rounded-md" role="button">Home</a>
+  </div>
+  <div>
+    <a href="{% url 'subscriber_manage_webhook_list' management_key=subscriber.management_key %}" class="bg-indigo-500 hover:bg-indigo-600 text-white shadow px-4 py-2 rounded-md" role="button">Webhooks</a>
+  </div>
+{% endblock %}
+
+{% block home_content %}
+<div class="flex flex-col space-y-8">
+  <div class="space-y-4">
+    <div class="flex flex-row justify-between">
+      <div class="text-2xl">Webhooks</div>
+    </div>
+    <div>
+      <form action="" method="post">
+        {% csrf_token %}
+        <div class="flex flex-col space-y-2">
+          <div>Are you sure, you want to delete the Webhook with the URL <span class="font-bold">{{ webhook.payload_url }}</span>?</div>
+          <button type="submit" class="bg-indigo-500 hover:bg-indigo-600 text-white shadow px-4 py-2 rounded-md">Delete Webhook</button>
+        </div>
+      </form>
+    </div>
+  </div>
+</div>
+{% endblock %}
